requirements.txt
streamlit>=1.30,<2.0
letta-client>=1.0,<2.0
python-dotenv>=1.0,<2.0
pymongo>=4.6,<5.0
